EU-115 transfer get_current_user to dependencies
